Byron only testnet and chairman test
